refactor(activity): keep bridge history lookup in activity selectors
Move selectBridgeHistoryItemForTx out of bridge-status and replace
useBridgeTxHistoryData in pending activity actions with the activity selector.
