+++ ui/pages/activity/rows/pending-transaction-actions.test.tsx
@@ -1,17 +1,26 @@
 import React from 'react';
 import { render } from '@testing-library/react';
+import { useSelector } from 'react-redux';
 import type { TransactionGroup } from '../../../../shared/lib/multichain/types';
 import { usePendingTransactionActions } from '../../../hooks/usePendingTransactionActions';
-import { useBridgeTxHistoryData } from '../../../hooks/bridge/useBridgeTxHistoryData';
+import { selectBridgeHistoryItemForTx } from '../../../selectors/activity';
 import { TransactionListItemPendingActions } from './pending-transaction-actions';
 
+jest.mock('react-redux', () => ({
+  useSelector: jest.fn(),
+}));
 jest.mock('../../../hooks/usePendingTransactionActions');
-jest.mock('../../../hooks/bridge/useBridgeTxHistoryData');
+jest.mock('../../../selectors/activity', () => ({
+  selectBridgeHistoryItemForTx: jest.fn(),
+}));
 
 const mockUsePendingTransactionActions = jest.mocked(
   usePendingTransactionActions,
 );
-const mockUseBridgeTxHistoryData = jest.mocked(useBridgeTxHistoryData);
+const mockUseSelector = jest.mocked(useSelector);
+const mockSelectBridgeHistoryItemForTx = jest.mocked(
+  selectBridgeHistoryItemForTx,
+);
 
 let capturedProps: {
   onCancel: (event: React.MouseEvent) => void;
@@ -65,9 +74,8 @@ describe('TransactionListItemPendingActions', () => {
   beforeEach(() => {
     capturedProps = null;
     jest.clearAllMocks();
-    mockUseBridgeTxHistoryData.mockReturnValue({
-      bridgeHistoryItem: undefined,
-    });
+    mockSelectBridgeHistoryItemForTx.mockReturnValue(undefined);
+    mockUseSelector.mockImplementation((selector) => selector({} as never));
     mockActions();
   });
