+++ ui/pages/activity/rows/pending-transaction-actions.tsx
@@ -1,11 +1,13 @@
 import React, { type MouseEvent as ReactMouseEvent } from 'react';
+import { useSelector } from 'react-redux';
 import type { TransactionMeta } from '@metamask/transaction-controller';
 import type { EditGasModes } from '../../../../shared/constants/gas';
 import type { TransactionGroup } from '../../../../shared/lib/multichain/types';
 import { isIntentBridgeActivity } from '../../../helpers/transactions/pending-transaction-actions';
-import { useBridgeTxHistoryData } from '../../../hooks/bridge/useBridgeTxHistoryData';
 import { usePendingTransactionActions } from '../../../hooks/usePendingTransactionActions';
 import { PendingTransactionActionButtons } from '../../../components/app/pending-transaction-action-buttons/pending-transaction-action-buttons';
+import { selectBridgeHistoryItemForTx } from '../../../selectors/activity';
+import type { MetaMaskReduxState } from '../../../store/store';
 
 type TransactionMetaWithSmartTransaction = TransactionMeta & {
   isSmartTransaction?: boolean;
@@ -25,7 +27,9 @@ export const TransactionListItemPendingActions = ({
   onGasModalMetaId,
 }: Readonly<TransactionListItemPendingActionsProps>) => {
   const { primaryTransaction, initialTransaction } = transactionGroup;
-  const { bridgeHistoryItem } = useBridgeTxHistoryData({ transactionGroup });
+  const bridgeHistoryItem = useSelector((state: MetaMaskReduxState) =>
+    selectBridgeHistoryItemForTx(state, initialTransaction),
+  );
   const { showCancel, onCancel, speedUp } = usePendingTransactionActions({
     transactionGroup,
     isEarliestNonce,
